+++ src/ui/MainMenu.hpp
@@ -3,7 +3,7 @@
 #include "../constants.hpp"
 #include "../ecs/ECSManager.hpp"
 #include "../engine/Engine.hpp"
-#include "../engine/Vec2d.hpp"
+#include "../engine/Vec2i.hpp"
 #include "../modules/GameStateManager.hpp"
 #include "../modules/MenuStack.hpp"
 #include "../modules/SaveGameManager.hpp"
@@ -26,7 +26,7 @@ class MainMenu final : public ListDialog {
 		                   menuStack.pop();
 	                   }},
 	                  {"EXIT", [&game]() { game.stop(); }}},
-	                 Vec2d{x, y}, menuWidth_)
+	                 Vec2i{x, y}, menuWidth_)
 	{
 	}
